+++ pkg/sql/colexec/aggregator.go
@@ -266,7 +266,7 @@ func makeAggregateFuncs(
 	return funcs, outTyps, nil
 }
 
-func (a *orderedAggregator) initWithOutputBatchSize(outputSize uint16) {
+func (a *orderedAggregator) initWithOutputBatchSize(outputSize uint64) {
 	a.initWithInputAndOutputBatchSize(int(coldata.BatchSize()), int(outputSize))
 }
 
@@ -329,7 +329,7 @@ func (a *orderedAggregator) Next(ctx context.Context) coldata.Batch {
 		a.scratch.resumeIdx = newResumeIdx
 		if a.scratch.resumeIdx >= a.scratch.outputSize {
 			// We still have overflow output values.
-			a.scratch.SetLength(uint16(a.scratch.outputSize))
+			a.scratch.SetLength(uint64(a.scratch.outputSize))
 			a.allocator.PerformOperation(a.unsafeBatch.ColVecs(), func() {
 				for i := 0; i < len(a.outputTypes); i++ {
 					a.unsafeBatch.ColVec(i).Copy(
@@ -338,7 +338,7 @@ func (a *orderedAggregator) Next(ctx context.Context) coldata.Batch {
 								Src:         a.scratch.ColVec(i),
 								ColType:     a.outputTypes[i],
 								SrcStartIdx: 0,
-								SrcEndIdx:   uint64(a.scratch.Length()),
+								SrcEndIdx:   a.scratch.Length(),
 							},
 						},
 					)
@@ -384,7 +384,7 @@ func (a *orderedAggregator) Next(ctx context.Context) coldata.Batch {
 
 	batchToReturn := a.scratch.Batch
 	if a.scratch.resumeIdx > a.scratch.outputSize {
-		a.scratch.SetLength(uint16(a.scratch.outputSize))
+		a.scratch.SetLength(uint64(a.scratch.outputSize))
 		a.allocator.PerformOperation(a.unsafeBatch.ColVecs(), func() {
 			for i := 0; i < len(a.outputTypes); i++ {
 				a.unsafeBatch.ColVec(i).Copy(
@@ -393,7 +393,7 @@ func (a *orderedAggregator) Next(ctx context.Context) coldata.Batch {
 							Src:         a.scratch.ColVec(i),
 							ColType:     a.outputTypes[i],
 							SrcStartIdx: 0,
-							SrcEndIdx:   uint64(a.scratch.Length()),
+							SrcEndIdx:   a.scratch.Length(),
 						},
 					},
 				)
@@ -403,7 +403,7 @@ func (a *orderedAggregator) Next(ctx context.Context) coldata.Batch {
 		batchToReturn = a.unsafeBatch
 		a.scratch.shouldResetInternalBatch = false
 	} else {
-		a.scratch.SetLength(uint16(a.scratch.resumeIdx))
+		a.scratch.SetLength(uint64(a.scratch.resumeIdx))
 		a.scratch.shouldResetInternalBatch = true
 	}
