+++ pkg/sql/colexec/min_max_agg_tmpl.go
@@ -132,7 +132,7 @@ func (a *_AGG_TYPEAgg) CurrentOutputIndex() int {
 func (a *_AGG_TYPEAgg) SetOutputIndex(idx int) {
 	if a.curIdx != -1 {
 		a.curIdx = idx
-		a.nulls.UnsetNullsAfter(uint16(idx + 1))
+		a.nulls.UnsetNullsAfter(uint64(idx + 1))
 	}
 }
 
@@ -146,7 +146,7 @@ func (a *_AGG_TYPEAgg) Compute(b coldata.Batch, inputIdxs []uint32) {
 		// any non-nulls for this group so far, the output for this group should
 		// be null.
 		if !a.foundNonNullForCurrentGroup {
-			a.nulls.SetNull(uint16(a.curIdx))
+			a.nulls.SetNull(uint64(a.curIdx))
 		} else {
 			a.allocator.PerformOperation(
 				[]coldata.Vec{a.vec},
@@ -214,7 +214,7 @@ func _ACCUMULATE_MINMAX(a *_AGG_TYPEAgg, nulls *coldata.Nulls, i int, _HAS_NULLS
 		// negative, it means that this is the first group.
 		if a.curIdx >= 0 {
 			if !a.foundNonNullForCurrentGroup {
-				a.nulls.SetNull(uint16(a.curIdx))
+				a.nulls.SetNull(uint64(a.curIdx))
 			} else {
 				execgen.SET(a.col, a.curIdx, a.curAgg)
 			}
@@ -224,7 +224,7 @@ func _ACCUMULATE_MINMAX(a *_AGG_TYPEAgg, nulls *coldata.Nulls, i int, _HAS_NULLS
 	}
 	var isNull bool
 	// {{ if .HasNulls }}
-	isNull = nulls.NullAt(uint16(i))
+	isNull = nulls.NullAt(uint64(i))
 	// {{ else }}
 	isNull = false
 	// {{ end }}
