+++ pkg/sql/colexec/types_integration_test.go
@@ -59,7 +59,7 @@ func TestSupportedSQLTypesIntegration(t *testing.T) {
 	rng, _ := randutil.NewPseudoRand()
 
 	for _, typ := range allSupportedSQLTypes {
-		for _, numRows := range []uint16{
+		for _, numRows := range []uint64{
 			// A few interesting sizes.
 			1,
 			coldata.BatchSize() - 1,
@@ -67,7 +67,7 @@ func TestSupportedSQLTypesIntegration(t *testing.T) {
 			coldata.BatchSize() + 1,
 		} {
 			rows := make(sqlbase.EncDatumRows, numRows)
-			for i := uint16(0); i < numRows; i++ {
+			for i := uint64(0); i < numRows; i++ {
 				rows[i] = make(sqlbase.EncDatumRow, 1)
 				rows[i][0] = sqlbase.DatumToEncDatum(&typ, sqlbase.RandDatum(rng, &typ, true /* nullOk */))
 			}
